refactor: extract queue drag-and-drop logic into a reusable QueueDragMixin and clean up redundant sortable utilities
